# Student Management System/gen/models/student.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Student student
//
// swagger:model student
type Student struct {

	// age
	// Example: 25
	Age string `json:"age,omitempty"`

	// city
	// Example: Bahawalpur
	City string `json:"city,omitempty"`

	// class
	// Example: BSCS
	Class string `json:"class,omitempty"`

	// email
	// Example: abdulmueez@wanclouds.net
	// Required: true
	// Format: email
	Email *strfmt.Email `json:"email"`

	// name
	// Example: Mueez Khan
	Name string `json:"name,omitempty"`

	// subject
	// Example: Golang
	Subject string `json:"subject,omitempty"`
}

// Validate validates this student
func (m *Student) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateEmail(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Student) validateEmail(formats strfmt.Registry) error {

	if err := validate.Required("email", "body", m.Email); err != nil {
		return err
	}

	if err := validate.FormatOf("email", "body", "email", m.Email.String(), formats); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this student based on context it is used
func (m *Student) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Student) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Student) UnmarshalBinary(b []byte) error {
	var res Student
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}