# Student Management System/gen/restapi/operations/student/createstudent_responses.go
// Code generated by go-swagger; DO NOT EDIT.

package student

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"net/http"

	"github.com/go-openapi/runtime"
)

// CreatestudentBadRequestCode is the HTTP code returned for type CreatestudentBadRequest
const CreatestudentBadRequestCode int = 400

/*CreatestudentBadRequest Invalid email exisit

swagger:response createstudentBadRequest
*/
type CreatestudentBadRequest struct {
}

// NewCreatestudentBadRequest creates CreatestudentBadRequest with default headers values
func NewCreatestudentBadRequest() *CreatestudentBadRequest {

	return &CreatestudentBadRequest{}
}

// WriteResponse to the client
func (o *CreatestudentBadRequest) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(400)
}

/*CreatestudentDefault successful operation

swagger:response createstudentDefault
*/
type CreatestudentDefault struct {
	_statusCode int
}

// NewCreatestudentDefault creates CreatestudentDefault with default headers values
func NewCreatestudentDefault(code int) *CreatestudentDefault {
	if code <= 0 {
		code = 500
	}

	return &CreatestudentDefault{
		_statusCode: code,
	}
}

// WithStatusCode adds the status to the createstudent default response
func (o *CreatestudentDefault) WithStatusCode(code int) *CreatestudentDefault {
	o._statusCode = code
	return o
}

// SetStatusCode sets the status to the createstudent default response
func (o *CreatestudentDefault) SetStatusCode(code int) {
	o._statusCode = code
}

// WriteResponse to the client
func (o *CreatestudentDefault) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(o._statusCode)
}