# Student Management System/swagger.yaml
swagger: "2.0"
info:
  description: "This Student Management System UI for Assestment Task By WanClouds."
  version: "1.0.0"
  title: "SMS By GOlang"
  contact:
    email: "abdulmueez@wanclouds.net"
host: "sms.swagger.io"
basePath: "/v1"
...
schemes:
  - "http"
paths:
...
  /Student/newstudent/add:
    post:
      tags:
        - "Student"
      summary: "Add student recoard"
      operationId: "createstudent"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created student recoard"
          required: true
          schema:
            $ref: "#/definitions/student"
      responses:
        "400":
          description: "Invalid email exisit"
        default:
          description: "successful operation"
...
definitions:
...
  student:
    type: "object"
    required:
      - "email"
    properties:
      name:
        type: "string"
        example: "Mueez Khan"
      email:
        type: "string"
        format: "email"
        example: "abdulmueez@wanclouds.net"
      class:
        type: "string"
        example: "BSCS"
      age:
        type: "string"
        example: "25"
      city:
        type: "string"
        example: "Bahawalpur"
      subject:
        type: "string"
        example: "Golang"